requirements.txt
streamlit
reportlab
python-docx
requests
beautifulsoup4
spacy
python -m spacy download en_core_web_sm
